+++ Console Game/Game/Runtime/Enemy/Types/EnemyWithReward.cs
@@ -20,7 +20,7 @@ namespace Console_Game
 
         public IEnemyData Data => _enemy.Data;
 
-        public void Update(long deltaTime)
+        public void Update(float deltaTime)
         {
             if (_enemy.Health.IsDied() && _reward.IsApplied == false)
                 _reward.Apply();
