+++ Console Game/Game/Runtime/Game/Loop/Interfaces/IGameLoopObject.cs
@@ -2,6 +2,6 @@ namespace Console_Game
 {
     public interface IGameLoopObject
     {
-        void Update(long deltaTime);
+        void Update(float deltaTime);
     }
 }
